+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Practice Dummies (OpenMW)
 
+## 1.2
+
+- Added ability to level Marksman using TD and OAAB shooting targets
+- Added compatibility for Damage Based Skill Progression
+
 ## 1.1
 
 - Added Time Flies interop